Moving more abort logic to its own mod
