+++ apps/src/lib/node/ledger/mod.rs
@@ -508,116 +508,3 @@ async fn run_abci(
         .await
         .map_err(|err| Error::TowerServer(err.to_string()))
 }
-
-/// Function that blocks until either
-///   1. User sends a shutdown signal
-///   2. One of the child processes terminates, sending a message on `drop`
-/// Returns a boolean to indicate which scenario occurred.
-/// `true` means that the latter happened
-#[cfg(unix)]
-async fn wait_for_abort(
-    mut abort_recv: tokio::sync::mpsc::UnboundedReceiver<&'static str>,
-) -> bool {
-    use tokio::signal::unix::{signal, SignalKind};
-    let mut sigterm = signal(SignalKind::terminate()).unwrap();
-    let mut sighup = signal(SignalKind::hangup()).unwrap();
-    let mut sigpipe = signal(SignalKind::pipe()).unwrap();
-    tokio::select! {
-        signal = tokio::signal::ctrl_c() => {
-            match signal {
-                Ok(()) => tracing::info!("Received interrupt signal, exiting..."),
-                Err(err) => tracing::error!("Failed to listen for CTRL+C signal: {}", err),
-            }
-        },
-        signal = sigterm.recv() => {
-            match signal {
-                Some(()) => tracing::info!("Received termination signal, exiting..."),
-                None => tracing::error!("Termination signal cannot be caught anymore, exiting..."),
-            }
-        },
-        signal = sighup.recv() => {
-            match signal {
-                Some(()) => tracing::info!("Received hangup signal, exiting..."),
-                None => tracing::error!("Hangup signal cannot be caught anymore, exiting..."),
-            }
-        },
-        signal = sigpipe.recv() => {
-            match signal {
-                Some(()) => tracing::info!("Received pipe signal, exiting..."),
-                None => tracing::error!("Pipe signal cannot be caught anymore, exiting..."),
-            }
-        },
-        msg = abort_recv.recv() => {
-            // When the msg is `None`, there are no more abort senders, so both
-            // Tendermint and the shell must have already exited
-            if let Some(who) = msg {
-                 tracing::info!("{} has exited, shutting down...", who);
-            }
-            return true;
-        }
-    };
-    false
-}
-
-/// Function that blocks until either
-///   1. User sends a shutdown signal
-///   2. One of the child processes terminates, sending a message on `drop`
-/// Returns a boolean to indicate which scenario occurred.
-/// `true` means that the latter happened
-#[cfg(windows)]
-async fn wait_for_abort(
-    mut abort_recv: tokio::sync::mpsc::UnboundedReceiver<&'static str>,
-) -> bool {
-    let mut sigbreak = tokio::signal::windows::ctrl_break().unwrap();
-    let _ = tokio::select! {
-        signal = tokio::signal::ctrl_c() => {
-            match signal {
-                Ok(()) => tracing::info!("Received interrupt signal, exiting..."),
-                Err(err) => tracing::error!("Failed to listen for CTRL+C signal: {}", err),
-            }
-        },
-        signal = sigbreak.recv() => {
-            match signal {
-                Some(()) => tracing::info!("Received break signal, exiting..."),
-                None => tracing::error!("Break signal cannot be caught anymore, exiting..."),
-            }
-        },
-        msg = abort_recv.recv() => {
-            // When the msg is `None`, there are no more abort senders, so both
-            // Tendermint and the shell must have already exited
-            if let Some(who) = msg {
-                 tracing::info!("{} has exited, shutting down...", who);
-            }
-            return true;
-        }
-    };
-    false
-}
-
-/// Function that blocks until either
-///   1. User sends a shutdown signal
-///   2. One of the child processes terminates, sending a message on `drop`
-/// Returns a boolean to indicate which scenario occurred.
-/// `true` means that the latter happened
-#[cfg(not(any(unix, windows)))]
-async fn wait_for_abort(
-    mut abort_recv: tokio::sync::mpsc::UnboundedReceiver<&'static str>,
-) -> bool {
-    let _ = tokio::select! {
-        signal = tokio::signal::ctrl_c() => {
-            match signal {
-                Ok(()) => tracing::info!("Received interrupt signal, exiting..."),
-                Err(err) => tracing::error!("Failed to listen for CTRL+C signal: {}", err),
-            }
-        },
-        msg = abort_recv.recv() => {
-            // When the msg is `None`, there are no more abort senders, so both
-            // Tendermint and the shell must have already exited
-            if let Some(who) = msg {
-                 tracing::info!("{} has exited, shutting down...", who);
-            }
-            return true;
-        }
-    };
-    false
-}
